enable Visual Studio toolset for CMake generator
Visual Studio provides, depending on the installed components and
versions, different compiler toolsets. These toolsets define, among
other things, the versions of the actual compiler, standard
library and runtime. To CMake, the toolset is provided via the `-T`
command line parameter. Frequent usage of toolsets in Visual Studio is
to target Windows XP (e.g. toolset `v120_xp` or `v140_xp`) or to use the
experimental Clang compiler frontend (`v140_clang_c2`).

In the future, this may be extended to support Xcode 3.0 and above,
which also has the notion of compiler toolsets.

Signed-off-by: Torbjörn Klatt <t.klatt.oss@mailbox.org>